BackendFeatures/DeltaLakeIntegration/requirements.txt
boto3
aws-cdk-lib==2.128.0
constructs>=10.0.0,<11.0.0
